+++ src/main/java/shef/data/Recipe.java
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+package shef.data;
+
 import java.util.LinkedList;
 import java.util.List;
 import java.util.logging.*;
