+++ src/main/java/shef/data/SpinOff.java
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+package shef.data;
+
 public class SpinOff extends Recipe {
 
   private Recipe original;
