+++ src/main/java/shef/data/UserComment.java
@@ -12,21 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-public class Step {
+package shef.data;
+ 
+/** A user's comment, with corresponding user info. */
+public class UserComment {
+  long id;
+  String comment;
+  long timestamp;
 
-  private String instruction;
-
-  public Step(String instruction) {
-    this.instruction = instruction;
+ /**
+  * @param id The entity's id.
+  * @param comment The user's comment.
+  * @param timestamp The time at which the comment was submitted.
+  */
+  public UserComment(long id, String comment, long timestamp) {
+      this.id = id;
+      this.comment = comment;
+      this.timestamp = timestamp;
   }
-
-  public String getInstruction() {
-    return instruction;
-  }
-
-  @Override
-  public String toString() {
-    return instruction;
-  }
-
-}
+}
