+++ src/main/java/shef/servlets/DisplayCommentsServlet.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.sps.servlets;
+package shef.servlets;
 
 import com.google.sps.data.UserComment;
 import com.google.appengine.api.datastore.DatastoreService;
